pyproject.toml
[project]
name = "handwritingsanalysis"
version = "0.1.0"
description = "extracting the hadnwritten words"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "numpy>=2.3.4",
    "ollama>=0.1.0",
    "opencv-python>=4.11.0.86",
    "paddlepaddle-gpu>=2.5.0",
    "paddleocr>=3.3.1",
    "pillow>=12.0.0",
    "svlearn-bootcamp>=0.1.7",
    "torch>=2.9.0",
    "transformers>=4.30.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "faster-whisper>=1.2.1",
    "librosa>=0.11.0",
    "pypdf2>=3.0.1",
    "spacy>=3.8.7",
    "networkx>=3.5",
    "matplotlib>=3.10.7",
    "wordcloud>=1.9.4",
    "en-core-web-sm",
    "config>=0.5.1",
    "prompt>=0.4.1",
    "speechrecognition>=3.14.3",
    "pyttsx3>=2.99",
    "scipy>=1.16.3",
    "pyaudio>=0.2.14",
    "streamlit-audiorec>=0.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/handwrittenanalysis"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz" }
